test(PawnStrategy): 메서드명 변경에 따른 변화
